requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
dnspython>=2.0.0
dkimpy>=1.1.0
